requirements.txt
Flask
requests
python-dotenv
flask_SQLAlchemy
flask_login
psycopg2